+++ client/tc/HISTORY.rst
@@ -3,6 +3,14 @@
 Release History
 ===============
 
+0.4.0
+++++++
++ Renamed default tc-base-url -> tc-url
+* Added org default tc-org
+* Added support for dynamic input and prompting
+* Fixed several bugs
+* Fixed several linting issues
+
 0.3.3
 ++++++
 * Preview release